+++ crates/test-issuer/src/lib.rs
@@ -8,8 +8,8 @@ use anyhow::anyhow;
 use chrono::{DateTime, Utc};
 use credibil_did::{DidResolver, Document};
 use credibil_infosec::{Algorithm, Signer};
-use credibil_vc::oid4vci::issuer::{Client, Dataset, Issuer, Server};
 use credibil_vc::oid4vci::provider::{Metadata, Provider, Subject};
+use credibil_vc::oid4vci::types::{Client, Dataset, Issuer, Server};
 use credibil_vc::openid::provider::{Result, StateStore};
 use credibil_vc::status::issuer::Status;
 pub use keystore::Keystore;
