+++ src/oid4vci/handlers/authorize.rs
@@ -72,12 +72,12 @@ use tracing::instrument;
 
 use crate::core::generate;
 use crate::oid4vci::endpoint::Request;
-use crate::oid4vci::issuer::{
+use crate::oid4vci::provider::{Metadata, Provider, Subject};
+use crate::oid4vci::state::{Authorization, AuthorizedItem, Expire, ItemType, Stage, State};
+use crate::oid4vci::types::{
     AuthorizationDetail, AuthorizationDetailType, AuthorizationRequest, AuthorizationResponse,
     Claim, CredentialAuthorization, Issuer, ProfileClaims, RequestObject,
 };
-use crate::oid4vci::provider::{Metadata, Provider, Subject};
-use crate::oid4vci::state::{Authorization, AuthorizedItem, Expire, ItemType, Stage, State};
 use crate::oid4vci::{Error, Result};
 use crate::openid::oauth::GrantType;
 use crate::openid::provider::StateStore;
