+++ src/oid4vci/handlers/credential_offer.rs
@@ -17,9 +17,9 @@
 use tracing::instrument;
 
 use crate::oid4vci::endpoint::Request;
-use crate::oid4vci::issuer::{CredentialOfferRequest, CredentialOfferResponse};
 use crate::oid4vci::provider::Provider;
 use crate::oid4vci::state::{Stage, State};
+use crate::oid4vci::types::{CredentialOfferRequest, CredentialOfferResponse};
 use crate::oid4vci::{Error, Result};
 use crate::openid::provider::StateStore;
