+++ src/oid4vci/handlers/par.rs
@@ -12,9 +12,9 @@ use tracing::instrument;
 use crate::core::generate;
 use crate::oid4vci::endpoint::Request;
 use crate::oid4vci::handlers::authorize;
-use crate::oid4vci::issuer::{PushedAuthorizationRequest, PushedAuthorizationResponse};
 use crate::oid4vci::provider::{Metadata, Provider};
 use crate::oid4vci::state::{PushedAuthorization, Stage, State};
+use crate::oid4vci::types::{PushedAuthorizationRequest, PushedAuthorizationResponse};
 use crate::oid4vci::{Error, Result};
 use crate::openid::provider::StateStore;
