+++ src/oid4vci/handlers/register.rs
@@ -3,9 +3,9 @@
 use tracing::instrument;
 
 use crate::oid4vci::endpoint::Request;
-use crate::oid4vci::issuer::{RegistrationRequest, RegistrationResponse};
 use crate::oid4vci::provider::Provider;
 use crate::oid4vci::state::State;
+use crate::oid4vci::types::{RegistrationRequest, RegistrationResponse};
 use crate::oid4vci::{Error, Result};
 use crate::openid::provider::StateStore;
