+++ src/openid.rs
@@ -3,7 +3,5 @@
 //! Types and logic used in the `OpenID4VC` specifications and consumed by
 //! `issuer` and `verifier` feature endpoints.
 
-
 pub mod oauth;
 pub mod provider;
-
