+++ tests/vci_register.rs
@@ -5,7 +5,7 @@ mod utils;
 use chrono::Utc;
 use credibil_vc::oid4vci;
 use credibil_vc::oid4vci::state::{Expire, Stage, State, Token};
-use credibil_vc::oid4vci::issuer::RegistrationRequest;
+use credibil_vc::oid4vci::types::RegistrationRequest;
 use credibil_vc::openid::provider::StateStore;
 use insta::assert_yaml_snapshot as assert_snapshot;
 use serde_json::json;
